refactor: move data formatting to [tab] static props
