+++ src/pages/teachers/curriculum/[subjectPhaseSlug]/[tab].tsx
@@ -4,7 +4,7 @@ import {
   GetStaticPropsResult,
   NextPage,
 } from "next";
-import React from "react";
+import React, { MutableRefObject } from "react";
 import { useRouter } from "next/router";
 
 import CMSClient from "@/node-lib/cms";
@@ -29,11 +29,47 @@ import { fetchSubjectPhasePickerData } from "@/pages/teachers/curriculum/index";
 import getPageProps from "@/node-lib/getPageProps";
 import OakError from "@/errors/OakError";
 import { buildCurriculumMetadata } from "@/components/CurriculumComponents/helpers/curriculumMetadata";
+import {
+  Thread,
+  Subject,
+  Domain,
+  Discipline,
+  Tier,
+  Unit,
+} from "@/components/CurriculumComponents/CurriculumVisualiser";
+import { YearSelection } from "@/components/CurriculumComponents/UnitsTab/UnitsTab";
 
 export type CurriculumSelectionSlugs = {
   phaseSlug: string;
   subjectSlug: string;
   examboardSlug: string | null;
+};
+
+export type CurriculumUnitsYearGroup = {
+  units: Unit[];
+  childSubjects: Subject[];
+  domains: Domain[];
+  tiers: Tier[];
+  disciplines: Discipline[];
+  ref?: MutableRefObject<HTMLDivElement>;
+};
+
+export type CurriculumUnitsYearData = {
+  [key: string]: {
+    units: Unit[];
+    childSubjects: Subject[];
+    domains: Domain[];
+    tiers: Tier[];
+    disciplines: Discipline[];
+    ref?: MutableRefObject<HTMLDivElement>;
+  };
+};
+
+export type CurriculumUnitsFormattedData = {
+  yearData: CurriculumUnitsYearData;
+  threadOptions: Thread[];
+  yearOptions: string[];
+  initialYearSelection: YearSelection;
 };
 
 export type CurriculumInfoPageProps = {
@@ -42,6 +78,7 @@ export type CurriculumInfoPageProps = {
   curriculumOverviewTabData: CurriculumOverviewMVData;
   curriculumOverviewSanityData: CurriculumOverviewSanityData;
   curriculumUnitsTabData: CurriculumUnitsTabData;
+  curriculumUnitsFormattedData: CurriculumUnitsFormattedData;
 };
 
 const VALID_TABS = ["overview", "units"] as const;
@@ -53,6 +90,7 @@ const CurriculumInfoPage: NextPage<CurriculumInfoPageProps> = ({
   curriculumOverviewTabData,
   curriculumUnitsTabData,
   curriculumOverviewSanityData,
+  curriculumUnitsFormattedData,
 }) => {
   const router = useRouter();
   const tab = router.query.tab as CurriculumTab;
@@ -88,7 +126,11 @@ const CurriculumInfoPage: NextPage<CurriculumInfoPageProps> = ({
       break;
     case "units":
       tabContent = (
-        <UnitsTab data={curriculumUnitsTabData} examboardSlug={examboardSlug} />
+        <UnitsTab
+          data={curriculumUnitsTabData}
+          examboardSlug={examboardSlug}
+          formattedData={curriculumUnitsFormattedData}
+        />
       );
       break;
     default:
@@ -171,6 +213,178 @@ export const parseSubjectPhaseSlug = (slug: string) => {
   };
 };
 
+function formatCurriculumUnitsData(
+  data: CurriculumUnitsTabData,
+): CurriculumUnitsFormattedData {
+  const threadOptions = [] as Thread[];
+  const yearOptions = [] as string[];
+  const yearData = {} as CurriculumUnitsYearData;
+  const initialYearSelection = {} as YearSelection;
+
+  const unitSlugs = new Set<string>();
+  const duplicateUnitSlugs = new Set<string>();
+
+  unitSlugs.clear();
+  duplicateUnitSlugs.clear();
+
+  data.units.forEach((unit: Unit) => {
+    // Populate years object
+
+    if (yearOptions.every((yo) => yo !== unit.year)) {
+      yearOptions.push(unit.year);
+    }
+
+    // Populate threads object
+
+    unit.threads.forEach((thread) => {
+      if (threadOptions.every((to) => to.slug !== thread.slug)) {
+        threadOptions.push(thread);
+      }
+    });
+
+    // Check if the yearData object has an entry for the unit's year
+    // If not, initialize it with default values
+
+    let currentYearData = yearData[unit.year];
+    if (!currentYearData) {
+      currentYearData = {
+        units: [],
+        childSubjects: [],
+        domains: [],
+        tiers: [],
+        disciplines: [],
+      };
+      yearData[unit.year] = currentYearData;
+    }
+
+    // Add the current unit
+
+    currentYearData.units.push(unit);
+
+    // Populate list of child subject filter values
+    if (
+      unit.subject_parent &&
+      unit.subject_parent_slug &&
+      currentYearData.childSubjects.every(
+        (c) => c.subject_slug !== unit.subject_slug,
+      )
+    ) {
+      currentYearData.childSubjects.push({
+        subject: unit.subject,
+        subject_slug: unit.subject_slug,
+      });
+    }
+
+    // Populate list of domain filter values
+
+    if (
+      unit.domain &&
+      unit.domain_id &&
+      currentYearData.domains.every((d) => d.domain_id !== unit.domain_id)
+    ) {
+      currentYearData.domains.push({
+        domain: unit.domain,
+        domain_id: unit.domain_id,
+      });
+    }
+
+    // Populate list of tier filter values
+
+    if (
+      unit.tier &&
+      unit.tier_slug &&
+      currentYearData.tiers.every((t) => t.tier_slug !== unit.tier_slug)
+    ) {
+      currentYearData.tiers.push({
+        tier: unit.tier,
+        tier_slug: unit.tier_slug,
+      });
+    }
+
+    // Loop through tags array and populate disciplines.
+    unit.tags?.forEach((tag) => {
+      if (tag.category === "Discipline") {
+        if (
+          currentYearData?.disciplines.findIndex((d) => d.id === tag.id) === -1
+        ) {
+          currentYearData.disciplines.push({
+            id: tag.id,
+            title: tag.title,
+          });
+        }
+      }
+    });
+
+    // Check for duplicate unit slugs
+
+    if (unitSlugs.has(unit.slug)) {
+      duplicateUnitSlugs.add(unit.slug);
+    } else {
+      unitSlugs.add(unit.slug);
+    }
+  });
+
+  // Sort year data
+
+  yearOptions.sort((a, b) => Number(a) - Number(b));
+
+  // Sort threads
+
+  const threadOrders = new Set(threadOptions.map((to) => to.order));
+  if (threadOptions.length > threadOrders.size) {
+    // In secondary science multiple threads can have the same order value due
+    // to multiple subjects (eg biology, chemistry, physics) being shown, so
+    // if orders are not unique, sort alphabetically by slug
+
+    threadOptions.sort((a, b) => a.slug.localeCompare(b.slug));
+  } else {
+    // If orders are unique, use them to sort
+
+    threadOptions.sort((a, b) => a.order - b.order);
+  }
+
+  Object.keys(yearData).forEach((year) => {
+    const filters = yearData[year];
+    if (!filters) {
+      throw new Error("year filters missing");
+    }
+    if (filters.domains.length > 0) {
+      filters.domains.sort((a, b) => a.domain_id - b.domain_id);
+      filters.domains.unshift({
+        domain: "All",
+        domain_id: 0,
+      });
+    }
+    filters.tiers.sort((a, b) => a.tier_slug.localeCompare(b.tier_slug));
+    // Sort disciplines
+    filters.disciplines.sort((a, b) => a.title.localeCompare(b.title));
+
+    // Add an "All" option if there are 2 or more disciplines. Set to -1 id as this shouldn't ever appear in the DB
+    const allDisciplineTag: Discipline = { id: -1, title: "All" };
+    if (filters.disciplines.length >= 2) {
+      filters.disciplines.unshift(allDisciplineTag);
+    }
+
+    initialYearSelection[year] = {
+      subject:
+        filters.childSubjects.find(
+          (s) => s.subject_slug === "combined-science",
+        ) ?? null,
+      discipline: allDisciplineTag,
+      domain: filters.domains.length ? filters.domains[0] : null,
+      tier: filters.tiers.length ? filters.tiers[0] : null,
+    };
+  });
+
+  const formattedDataCurriculumUnits = {
+    yearData,
+    threadOptions,
+    yearOptions,
+    initialYearSelection,
+  };
+  return formattedDataCurriculumUnits;
+}
+
 export const getStaticProps: GetStaticProps<
   CurriculumInfoPageProps,
   URLParams
@@ -209,6 +423,9 @@ export const getStaticProps: GetStaticProps<
       const curriculumUnitsTabData = await curriculumApi.curriculumUnits(slugs);
       curriculumUnitsTabData.units.sort((a, b) => a.order - b.order);
       const subjectPhaseOptions = await fetchSubjectPhasePickerData();
+      const curriculumUnitsFormattedData = formatCurriculumUnitsData(
+        curriculumUnitsTabData,
+      );
       const results: GetStaticPropsResult<CurriculumInfoPageProps> = {
         props: {
           curriculumSelectionSlugs: slugs,
@@ -216,6 +433,7 @@ export const getStaticProps: GetStaticProps<
           curriculumOverviewTabData,
           curriculumOverviewSanityData,
           curriculumUnitsTabData,
+          curriculumUnitsFormattedData,
         },
       };
       const resultsWithIsr = decorateWithIsr(results);
